Programming Exercise 2 finished!
I fixed the bugs that I had with the logic behind assigning the ticks within the axises inside of the graphs with the help of the professor!
I also finished the final part where it asks for the mean and finally, wrote my interpretation for each graph at the end of my code blocks!